backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.1
SQLAlchemy==2.0.23
Werkzeug==2.3.7
pytest==7.4.3
black==23.11.0
flake8==6.1.0
python-jose==3.3.0
bcrypt==4.0.1
requests==2.31.0
gunicorn==21.2.0 